requirements.txt
fastapi
uvicorn
pydantic
fhir.resources
python-multipart
pytesseract
SpeechRecognition
pillow
pandas
hl7
motor
pymongo
python-dotenv
aiofiles
jellyfish
scikit-learn
Levenshtein
xgboost
requests
langchain
langchain-community
openai
chromadb
sentence-transformers
